Start of component file reorganisation
