+++ frontend/src/views/MyProgress/introduction/DiversificationProgress.vue
