+++ frontend/src/views/MyProgress/introduction/WasteProgress.vue
